move validateRegistryCommit into utils
