+++ typescript/infra/scripts/rebalancer/deploy-rebalancer.ts
@@ -1,6 +1,4 @@
 import { input } from '@inquirer/prompts';
-import chalk from 'chalk';
-import { execSync } from 'child_process';
 import path from 'path';
 
 import {
@@ -10,10 +8,10 @@ import {
   rootLogger,
 } from '@hyperlane-xyz/utils';
 
-import { getRegistry } from '../../config/registry.js';
 import { DeployEnvironment } from '../../src/config/environment.js';
 import { RebalancerHelmManager } from '../../src/rebalancer/helm.js';
 import { HelmCommand } from '../../src/utils/helm.js';
+import { validateRegistryCommit } from '../../utils.js';
 import {
   assertCorrectKubeContext,
   getArgs,
@@ -22,24 +20,6 @@ import {
   withWarpRouteId,
 } from '../agent-utils.js';
 import { getEnvironmentConfig } from '../core-utils.js';
-
-async function validateRegistryCommit(commit: string) {
-  const registry = getRegistry();
-  const registryUri = registry.getUri();
-
-  try {
-    rootLogger.info(
-      chalk.grey.italic(`Attempting to fetch registry commit ${commit}...`),
-    );
-    execSync(`cd ${registryUri} && git fetch origin ${commit}`, {
-      stdio: 'inherit',
-    });
-    rootLogger.info(chalk.grey.italic('Fetch completed successfully.'));
-  } catch (_) {
-    rootLogger.error(chalk.red(`Unable to fetch registry commit ${commit}.`));
-    process.exit(1);
-  }
-}
 
 function getRebalancerConfigPathPrefix(environment: DeployEnvironment) {
   return `config/environments/${environment}/rebalancer/rebalancerConfigs`;
