+++ typescript/infra/scripts/warp-routes/deploy-warp-monitor.ts
@@ -1,6 +1,4 @@
 import { input } from '@inquirer/prompts';
-import chalk from 'chalk';
-import { execSync } from 'child_process';
 
 import {
   LogFormat,
@@ -10,9 +8,9 @@ import {
 } from '@hyperlane-xyz/utils';
 
 import { Contexts } from '../../config/contexts.js';
-import { getRegistry } from '../../config/registry.js';
 import { HelmCommand } from '../../src/utils/helm.js';
 import { WarpRouteMonitorHelmManager } from '../../src/warp/helm.js';
+import { validateRegistryCommit } from '../../utils.js';
 import {
   assertCorrectKubeContext,
   getAgentConfig,
@@ -21,24 +19,6 @@ import {
   withWarpRouteId,
 } from '../agent-utils.js';
 import { getEnvironmentConfig } from '../core-utils.js';
-
-async function validateRegistryCommit(commit: string) {
-  const registry = getRegistry();
-  const registryUri = registry.getUri();
-
-  try {
-    rootLogger.info(
-      chalk.grey.italic(`Attempting to fetch registry commit ${commit}...`),
-    );
-    execSync(`cd ${registryUri} && git fetch origin ${commit}`, {
-      stdio: 'inherit',
-    });
-    rootLogger.info(chalk.grey.italic('Fetch completed successfully.'));
-  } catch (_) {
-    rootLogger.error(chalk.red(`Unable to fetch registry commit ${commit}.`));
-    process.exit(1);
-  }
-}
 
 async function main() {
   configureRootLogger(LogFormat.Pretty, LogLevel.Info);
